Render scrollback in the terminal; pin viewport to the latest output
The wezterm `Screen::for_each_phys_line` walks scrollback + visible
rows, but `sync_grid_snapshot` was sizing the rope to `physical_rows`
and treating `cursor.y` as a buffer-line index. Result: with any
scrollback the rope grew but the view stayed clamped to the top, so
fresh shell output scrolled off-screen instead of sticking to the
bottom — the opposite of how every terminal works.

Iterate the full `scrollback_rows()`, map `cursor.y` (visible-row
index) into a buffer-line index, and after each rebuild anchor
`ScrollState.scroll_offset` to `max_scroll` whenever the user is
already at the bottom. Wheeling up into history disengages the
follow; wheeling back within one row re-engages it.

Gate the rope/style rebuild on `Term::current_seqno()` and the
visible-row count so 10k-line scrollbacks don't pay the O(N) cost
on every idle frame. Widened `cursor_row` to `u32` since a buffer-
line index can exceed `u16::MAX` once scrollback fills.

+++ crates/bevy_terminal/src/cursor.rs
@@ -16,7 +16,7 @@ use crate::types::TerminalGridSnapshot;
 #[reflect(Component, Default)]
 pub struct TerminalCursorBlink {
     pub last_change_secs: f64,
-    pub last_row: u16,
+    pub last_row: u32,
     pub last_col: u16,
 }
 
@@ -66,7 +66,7 @@ pub fn push_terminal_caret(
 
         let x_left = snapshot.cursor_col as f32 * font.char_width;
         overlays.rects.push(caret_overlay(
-            snapshot.cursor_row as u32,
+            snapshot.cursor_row,
             x_left,
             &cursor_settings,
             theme.cursor,
